requirements.txt
pillow~=11.2.1
colorama~=0.4.6